Tweaked The UI
Added: a code assistant plugin which allows you to break down an entire project (repo) of code from .cs files and export them into text files.

Why? The text files are broken down to 14000 characters each.

This is the maximum that any web AI chat will allow a user to paste into their input text boxes at one time.

So yes you'll have to go through each text file and copy and paste them to your favorite web AI chat bot.

I personally use this with MistralAI and DeepAI and it works because they are able to piece together every file logically and in order despite each file being cut off at 14000 characters.

Tweaked: the assistant a lot so that the RichText box will expand when a user maximizes or minimizes the application. This is more for better viewing but I may add features to this later (as I need them).